requirements.txt
# Lab Report Analytics System - Requirements
# Core dependencies optimized for cloud deployment

# Image processing (headless version for server deployment)
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# Database - Async MongoDB
motor>=3.3.0
pymongo>=4.6.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# HTTP client for Groq API
httpx>=0.26.0

# Data validation
pydantic>=2.5.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
email-validator>=2.0.0

# Utilities
python-dotenv>=1.0.0
gunicorn>=21.0.0

# Optional OCR (commented out for smaller deployment - can enable later)
# paddleocr==2.9.1
# paddlepaddle==3.0.0
# scikit-image==0.21.0
